+++ airline-project/src/main/java/app/services/BookingService.java
@@ -30,8 +30,9 @@ public class BookingService {
     private final BookingRepository bookingRepository;
     private final FlightSeatService flightSeatService;
     private final BookingMapper bookingMapper;
-    private final PassengerService passengerService;
-
+    @Lazy
+    @Autowired
+    private PassengerService passengerService;
     @Lazy
     @Autowired
     private TicketService ticketService;
@@ -133,7 +134,8 @@ public class BookingService {
         return bookingRepository.findById(id);
     }
 
-    public Optional<Booking> getBookingByFlightSeatId(Long flightSeatId) {
-        return bookingRepository.findByFlightSeatId(flightSeatId);
+    @Transactional
+    public void deleteBookingByPassengerId(Long id) {
+        bookingRepository.deleteBookingByPassengerId(id);
     }
 }