+++ airline-project/src/main/java/app/services/PassengerService.java
@@ -28,6 +28,7 @@ public class PassengerService {
     private final FlightSeatService flightSeatService;
     private final PassengerRepository passengerRepository;
     private final PassengerMapper passengerMapper;
+    private final BookingService bookingService;
 
     public List<PassengerDto> getAllPassengers() {
         return passengerMapper.toDtoList(passengerRepository.findAll());
@@ -111,6 +112,7 @@ public class PassengerService {
     public void deletePassenger(Long id) {
         flightSeatService.makeFlightSeatNotSold(ticketService.getFlightSeatIdsByPassengerId(id));
         ticketService.deleteTicketByPassengerId(id);
+        bookingService.deleteBookingByPassengerId(id);
         passengerRepository.deleteById(id);
     }